:truck: Rename index.setup.php to wizard.php
